src/steps/40_NCPs/40_NCPs_requirements.txt
r-base
r-remotes
r-codetools
gdal
proj
r-terra
r-raster
r-data.table
r-sf
natcap.invest
r-yaml
pyyaml
# # # Dependencies by NCP:
#
# #  CAR
# r-terra
# natcap.invest
#
# #  FF
# r-terra
# Recocrop https://github.com/cropmodels/Recocrop
# r-data.table
# r-sf
# gdal
# proj
# #openmpi
# #gcc
#
# #  HAB
# natcap.invest
#
# #  ID
# r-terra
#
# #  MED
# r-terra
# r-data.table
# gdal
# proj
# #openmpi
# #gcc
#
# #  NDR
# natcap.invest
#
# #  PC
# r-terra
# r-data.table
#
# #  POL
# natcap.invest
# r-terra
# r-raster
#
# #  SDR
# natcap.invest
#
# #  WY
# natcap.invest